+++ translation-demo/src/components/BrandHeader.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/utils';
+import { cn } from '@/utils/cn';
 
 type Props = {
   className?: string;
