+++ translation-demo/src/components/PublisherPanel.tsx
@@ -10,7 +10,7 @@ import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
 import { Label } from '@/components/ui/label';
 
-import { usePublisher } from './usePublisher';
+import { usePublisher } from '@/hooks/usePublisher';
 
 const statusLabel = {
   connected: 'Live',
