+++ translation-demo/src/pages/publish.tsx
@@ -1,4 +1,4 @@
-import { PublisherPanel } from '@/components/moq/PublisherPanel';
+import { PublisherPanel } from '@/components/PublisherPanel';
 
 function PublishPage() {
   return <PublisherPanel />;
