+++ translation-demo/src/pages/watch.tsx
@@ -1,9 +1,9 @@
 import { Play } from 'lucide-react';
-import { useEffect, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { useNavigate, useParams } from 'react-router';
 
-import { StreamView } from '@/components/moq/StreamView';
-import { useMoqStreamViewer } from '@/components/moq/useMoqStreamViewer';
+import { StreamView } from '@/components/StreamView';
+import { useMoqStreamViewer } from '@/hooks/useMoqStreamViewer';
 import { Button } from '@/components/ui/button';
 
 function WatchPage() {
@@ -15,13 +15,16 @@ function WatchPage() {
   const [started, setStarted] = useState(false);
   const { connection, connectionStatus, stream } = useMoqStreamViewer(name, started);
 
-  // Track whether a stream has ever appeared (so losing it means it ended).
-  const [hasSeenStream, setHasSeenStream] = useState(false);
-  useEffect(() => {
-    if (stream) {
-      setHasSeenStream(true);
-    }
-  }, [stream]);
+  // Track whether a stream has ever appeared (so losing it means it ended). A ref is enough:
+  // nothing renders off this directly — it's only read to decide `streamUnavailable`, and the
+  // transition that matters (stream going away) re-renders on its own. Reset whenever we're not
+  // started so each watch session begins clean.
+  const hasSeenStreamRef = useRef(false);
+  if (!started) {
+    hasSeenStreamRef.current = false;
+  } else if (stream) {
+    hasSeenStreamRef.current = true;
+  }
 
   // After starting, give the stream a short window to be announced before deciding it is
   // not available (covers links to streams that never started / don't exist).
@@ -37,11 +40,10 @@ function WatchPage() {
 
   // Unavailable if a stream we were watching ended, or none ever showed up after the grace.
   const streamUnavailable =
-    started && !stream && (hasSeenStream || (waited && connectionStatus === 'connected'));
+    started && !stream && (hasSeenStreamRef.current || (waited && connectionStatus === 'connected'));
 
   const disconnect = () => {
     setStarted(false);
-    setHasSeenStream(false);
   };
 
   return (
