+++ translation-demo/src/utils/types.ts
